requirements.txt
alembic==1.4.2
appdirs==1.4.3
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
pipenv==2018.11.26
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.17
virtualenv==20.0.18
virtualenv-clone==0.5.4
Werkzeug==1.0.1
